add digest.Register
