+++ digest/register.go
@@ -3,12 +3,10 @@ package digest
 import (
 	"errors"
 	"fmt"
+	"io"
 )
 
-var (
-	ErrUnknown = errors.New("unrecognized digest algorithm")
-	Defaults   = DefaultRegister()
-)
+var ErrUnknown = errors.New("unrecognized digest algorithm")
 
 // Register is an immutable container of Algs.
 type Register struct {
@@ -26,14 +24,47 @@ func NewRegister(algs ...Alg) Register {
 	return newR
 }
 
-// New returns the Alg for the given id or ErrUnknown if the algorithm is not
+// Get returns the Alg for the given id or ErrUnknown if the algorithm is not
 // present in the register.
-func (r Register) New(id string) (Alg, error) {
+func (r Register) Get(id string) (Alg, error) {
 	alg, ok := r.algs[id]
 	if !ok {
 		return nil, fmt.Errorf("%w: %q", ErrUnknown, id)
 	}
 	return alg, nil
+}
+
+// NewDigester returns a digester for the given id, which must an Alg registered
+// in r.
+func (r Register) NewDigester(id string) (Digester, error) {
+	alg, err := r.Get(id)
+	if err != nil {
+		return nil, err
+	}
+	return alg.Digester(), nil
+}
+
+// NewMultiRegsiter returns a MultiDigester using algs from the register an
+// error returned if any alg is not defined in the r or if len(algs) < 1.
+func (r Register) NewMultiDigester(algs ...string) (*MultiDigester, error) {
+	if len(algs) < 1 {
+		return nil, errors.New("initializing multi-digester without digest algorithms")
+	}
+	writers := make([]io.Writer, 0, len(algs))
+	rs := make(map[string]Digester, len(algs))
+	for _, algID := range algs {
+		r, err := r.NewDigester(algID)
+		if err != nil {
+			return nil, err
+		}
+		rs[algID] = r
+		writers = append(writers, r)
+	}
+
+	return &MultiDigester{
+		Writer:    io.MultiWriter(writers...),
+		digesters: rs,
+	}, nil
 }
 
 // Append returns a new Register that includes algs from r plus additional algs.
@@ -52,8 +83,8 @@ func (r Register) Append(algs ...Alg) Register {
 	return newR
 }
 
-// Names returns names of all Alg IDs in r.
-func (r Register) Names() []string {
+// IDs returns IDs of all Algs in r.
+func (r Register) IDs() []string {
 	names := make([]string, 0, len(r.algs))
 	for name := range r.algs {
 		names = append(names, name)
@@ -61,5 +92,6 @@ func (r Register) Names() []string {
 	return names
 }
 
-// DefaultRegister returns a new Register with default Alg constructors.
-func DefaultRegister() Register { return NewRegister(builtin...) }
+// DefaultRegister returns a new Register with built-in Algs (sha512, sha256,
+// sha1, md5, and blake2b).
+func DefaultRegister() Register { return builtinRegister }
